+++ modules/apps/frontend-js/frontend-js-web/src/main/resources/META-INF/resources/liferay/modal/Modal.js
@@ -76,7 +76,8 @@ const Modal = ({
 	const onButtonClick = ({formId, onClick, type}) => {
 		if (type === 'cancel') {
 			processClose();
-		} else if (url && type === 'submit') {
+		}
+		else if (url && type === 'submit') {
 			const iframe = document.querySelector('.liferay-modal iframe');
 
 			if (iframe) {
@@ -97,7 +98,8 @@ const Modal = ({
 					if (form) {
 						form.submit();
 					}
-				} else if (forms.length >= 1) {
+				}
+				else if (forms.length >= 1) {
 					forms[0].submit();
 				}
 			}
@@ -299,7 +301,8 @@ const openPortletModal = ({
 				if (titleTextElement) {
 					title = `${titleTextElement.outerHTML} - ${title}`;
 				}
-			} else {
+			}
+			else {
 				title = `${titleElement.textContent} - ${title}`;
 			}
 		}
@@ -395,7 +398,8 @@ const openSelectionModal = ({
 					}
 				);
 			}
-		} else {
+		}
+		else {
 			onSelect(selectedItem);
 
 			processCloseFn();
@@ -452,7 +456,8 @@ const openSelectionModal = ({
 					if (selectedDataSet.has(itemId)) {
 						itemElement.disabled = true;
 						itemElement.classList.add('disabled');
-					} else {
+					}
+					else {
 						itemElement.disabled = false;
 						itemElement.classList.remove('disabled');
 					}
