+++ modules/apps/frontend-js/frontend-js-web/src/main/resources/META-INF/resources/liferay/modal/components/SimpleInputModal.es.js
@@ -72,14 +72,16 @@ const SimpleInputModal = ({
 						setLoadingResponse(false);
 
 						handleFormError(responseContent);
-					} else {
+					}
+					else {
 						setVisible(false);
 
 						closeModal();
 
 						if (responseContent.redirectURL) {
 							navigate(responseContent.redirectURL);
-						} else {
+						}
+						else {
 							if (onFormSuccess) {
 								onFormSuccess({
 									...responseContent,
