+++ modules/apps/frontend-js/frontend-js-web/src/main/resources/META-INF/resources/liferay/util/navigate.ts
@@ -12,7 +12,7 @@
  * details.
  */
 
-import { Callback } from "../types";
+import {Callback} from '../types';
 
 /**
  * Performs navigation to the given url. If SPA is enabled, it will route the
